services/pdf-vector/requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
PyMuPDF>=1.23.0,<1.27
python-docx==0.8.11
Pillow>=10.2.0
gunicorn==21.2.0
python-dotenv==1.0.0
pdfservices-sdk==4.2.0
cairosvg==2.7.1
reportlab==4.0.4
svglib==1.5.1